+++ CHANGES.md
@@ -1,5 +1,8 @@
 Change history of go-restful
 =
+2013-11-13
+ - (api add) Response knows how many bytes are written to the response body.
+
 2013-10-29
  - (api add) RecoverHandler(handler RecoverHandleFunction) to change how panic recovery is handled. Default behavior is to log and return a stacktrace. This may be a security issue as it exposes sourcecode information.